README.md
Global Superstore Power BI Dashboard

This repository contains a Power BI dashboard analyzing the Global Superstore data. The dashboard provides insights into sales, profits, and product performance across different cities, regions, and categories.


## Power BI Dashboard

The Power BI dashboard file (PBIX) can be found here. It showcases visualizations and interactive reports for analyzing Global superstore dataset.
## Getting Started

To view the dashboard and explore the data, follow these steps:

1. Download the PBIX file from the [Releases](/releases) section.
2. Open the PBIX file in Power BI Desktop.
3. Connect the dashboard to the Global Superstore dataset.
4. Interact with the visuals to gain insights into sales, profits, and product performance.

## Dataset

The dataset used in this dashboard is available here, It includes sales data, profits, units sold, order priority, shipping mode, and product categories and subcategories for the Global Superstore.

## Dashboard Preview 
It includes the image of the dashboard in PNG format

##Dashboard Background
It includes the background image of the dashboard in JPEG format

Feel free to explore and contribute to this repository. Feedback and suggestions are always welcome!